Immigrate suggested-password code from changePasswordDisplay to ChangePasword
